refactor: moved get-form-values to ../helpers
