--- shared/fragments/modals/src/modal-edit-project/helpers/get-form-values.helper.ts
@@ -1,9 +1,0 @@
-import { Project }           from '@shared/data'
-import { FormProjectValues } from '@ui/form'
-
-export const getFormValues = (project?: Project): FormProjectValues => ({
-  name: project?.name ?? '',
-  description: project?.description ?? '',
-  tags: project?.tags || [],
-  avatar: project?.avatarUrl || '',
-})
