--- shared/fragments/modals/src/modal-edit-team/helpers/get-form-values.helper.ts
@@ -1,9 +1,0 @@
-import { Team }           from '@shared/data'
-import { FormTeamValues } from '@ui/form'
-
-export const getFormValues = (team?: Team): FormTeamValues => ({
-  name: team?.name ?? '',
-  description: team?.description ?? '',
-  tags: team?.tags || [],
-  avatar: team?.avatarUrl || '',
-})
