--- shared/fragments/modals/src/modal-edit-team/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './get-form-values.helper'
